split modal into own component
