+++ lit/src/components/ar-icon/ar-icon.ts
@@ -1,0 +1,33 @@
+import { html, LitElement } from 'lit';
+import { customElement } from 'lit/decorators.js';
+import { styles } from '../../styles';
+
+@customElement('ar-icon')
+export class ArIcon extends LitElement {
+    static styles = styles;
+
+    render() {
+        return html`
+            <!-- Purchased with royality free license on from https://thenounproject.com/term/ar/2300882/ -->
+            <svg
+                class="ar-icon"
+                xmlns="http://www.w3.org/2000/svg"
+                xmlns:xlink="http://www.w3.org/1999/xlink"
+                viewBox="0 0 24 24"
+                version="1.1"
+                x="0px"
+                y="0px">
+
+                <title>icon/augmented-reality</title>
+                <g stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
+                    <path
+                        class="ar-icon-path"
+                        fill="#ffffff"
+                        fill-rule="nonzero"
+                        d="M19.7190894,6.05052769 C19.814566,6.09698073 19.8960604,6.17408691 19.9472136,6.2763932 C19.9496524,6.28127083 19.952004,6.28616925 19.954269,6.29108671 C19.9592282,6.3018659 19.9638087,6.31283194 19.9679984,6.32396506 C19.9910663,6.3852123 20.0012988,6.44811502 20,6.51005871 L20,10 C20,10.2761424 19.7761424,10.5 19.5,10.5 C19.2238576,10.5 19,10.2761424 19,10 L19,7.30901699 L16.7236068,8.4472136 C16.4766175,8.57070822 16.176281,8.47059605 16.0527864,8.2236068 C15.9292918,7.97661755 16.029404,7.67628103 16.2763932,7.5527864 L18.381966,6.5 L16.2763932,5.4472136 C16.029404,5.32371897 15.9292918,5.02338245 16.0527864,4.7763932 C16.176281,4.52940395 16.4766175,4.42929178 16.7236068,4.5527864 L19.7190894,6.05052769 Z M20,16.4899413 C20.0015392,16.5633494 19.9868837,16.6381045 19.954269,16.7089133 C19.9297792,16.7621441 19.8960558,16.8108178 19.8545725,16.8525503 C19.8124725,16.8943746 19.7675846,16.9258776 19.7190894,16.9494723 L16.7236068,18.4472136 C16.4766175,18.5707082 16.176281,18.470596 16.0527864,18.2236068 C15.9292918,17.9766175 16.029404,17.676281 16.2763932,17.5527864 L18.381966,16.5 L16.2763932,15.4472136 C16.029404,15.323719 15.9292918,15.0233825 16.0527864,14.7763932 C16.176281,14.529404 16.4766175,14.4292918 16.7236068,14.5527864 L19,15.690983 L19,13 C19,12.7238576 19.2238576,12.5 19.5,12.5 C19.7761424,12.5 20,12.7238576 20,13 L20,16.4899413 Z M4.61803399,16.5 L6.7236068,17.5527864 C6.97059605,17.676281 7.07070822,17.9766175 6.9472136,18.2236068 C6.82371897,18.470596 6.52338245,18.5707082 6.2763932,18.4472136 L3.28091063,16.9494723 C3.18543397,16.9030193 3.10393955,16.8259131 3.0527864,16.7236068 C3.05034759,16.7187292 3.04799599,16.7138307 3.045731,16.7089133 C3.04077183,16.6981341 3.0361913,16.6871681 3.03200163,16.6760349 C3.0089337,16.6147877 2.99870119,16.551885 3,16.4899413 L3,13 C3,12.7238576 3.22385763,12.5 3.5,12.5 C3.77614237,12.5 4,12.7238576 4,13 L4,15.690983 L6.2763932,14.5527864 C6.52338245,14.4292918 6.82371897,14.529404 6.9472136,14.7763932 C7.07070822,15.0233825 6.97059605,15.323719 6.7236068,15.4472136 L4.61803399,16.5 Z M4,7.30901699 L4,10 C4,10.2761424 3.77614237,10.5 3.5,10.5 C3.22385763,10.5 3,10.2761424 3,10 L3,6.51005871 C2.99846081,6.43665055 3.01311634,6.3618955 3.045731,6.29108671 C3.07022081,6.23785594 3.10394418,6.18918216 3.14542751,6.14744967 C3.18752755,6.10562541 3.23241545,6.07412245 3.28091063,6.05052769 L6.2763932,4.5527864 C6.52338245,4.42929178 6.82371897,4.52940395 6.9472136,4.7763932 C7.07070822,5.02338245 6.97059605,5.32371897 6.7236068,5.4472136 L4.61803399,6.5 L6.7236068,7.5527864 C6.97059605,7.67628103 7.07070822,7.97661755 6.9472136,8.2236068 C6.82371897,8.47059605 6.52338245,8.57070822 6.2763932,8.4472136 L4,7.30901699 Z M11.2673964,2.0572848 C11.3368783,2.02070348 11.4160214,2 11.5,2 C11.5839786,2 11.6631217,2.02070348 11.7326036,2.0572848 L14.7236068,3.5527864 C14.970596,3.67628103 15.0707082,3.97661755 14.9472136,4.2236068 C14.823719,4.47059605 14.5233825,4.57070822 14.2763932,4.4472136 L12,3.30901699 L12,6 C12,6.27614237 11.7761424,6.5 11.5,6.5 C11.2238576,6.5 11,6.27614237 11,6 L11,3.30901699 L8.7236068,4.4472136 C8.47661755,4.57070822 8.17628103,4.47059605 8.0527864,4.2236068 C7.92929178,3.97661755 8.02940395,3.67628103 8.2763932,3.5527864 L11.2673964,2.0572848 Z M11.7326036,20.9427152 C11.6631217,20.9792965 11.5839786,21 11.5,21 C11.4160214,21 11.3368783,20.9792965 11.2673964,20.9427152 L8.2763932,19.4472136 C8.02940395,19.323719 7.92929178,19.0233825 8.0527864,18.7763932 C8.17628103,18.529404 8.47661755,18.4292918 8.7236068,18.5527864 L11,19.690983 L11,17 C11,16.7238576 11.2238576,16.5 11.5,16.5 C11.7761424,16.5 12,16.7238576 12,17 L12,19.690983 L14.2763932,18.5527864 C14.5233825,18.4292918 14.823719,18.529404 14.9472136,18.7763932 C15.0707082,19.0233825 14.970596,19.323719 14.7236068,19.4472136 L11.7326036,20.9427152 Z M12,10.809017 L12,13.5 C12,13.7761424 11.7761424,14 11.5,14 C11.2238576,14 11,13.7761424 11,13.5 L11,10.809017 L8.2763932,9.4472136 C8.02940395,9.32371897 7.92929178,9.02338245 8.0527864,8.7763932 C8.17628103,8.52940395 8.47661755,8.42929178 8.7236068,8.5527864 L11.5,9.94098301 L14.2763932,8.5527864 C14.5233825,8.42929178 14.823719,8.52940395 14.9472136,8.7763932 C15.0707082,9.02338245 14.970596,9.32371897 14.7236068,9.4472136 L12,10.809017 Z">
+                    </path>
+                </g>
+            </svg>
+        `;
+    }
+}
