--- lit/src/components/ar-modal.ts
@@ -1,11 +1,0 @@
-import { html, LitElement } from 'lit';
-import { customElement } from 'lit/decorators.js';
-
-@customElement('ar-modal')
-export class ArModal extends LitElement {
-    render() {
-        return html`
-            I am ar modal you poopy butthole. <br>
-        `;
-    }
-}
